- ProgressDialog can now be hidden - Entering preferences during refresh of releases is impossible now (in order to avoid changes to preferences concerning refresh).
Changed LoadNewReleasesServiceTask to a POJO called
LoadNewReleasesServiceBinding that keeps a binding to the service and
updates the ProgressDialog. Why? The async task wasn't needed any more
at this point, as the service runs the logic in a separate thread
anyway. The only reason for keeping the async task was updates of the
GUI, which is now done via runOnUiThread().